// src/app/shared/models/Menu.ts
export interface Menu {
  dayCreate?: boolean;
  dayUpdate?: boolean;
  dayList?: boolean;

  hourCreate?: boolean;
  hourUpdate?: boolean;
  hourList?: boolean;

  donationCreate?: boolean;
  donationUpdate?: boolean;
  donationList?: boolean;

  employeeCreate?: boolean;
  employeeUpdate?: boolean;
  employeeList?: boolean;

  userCreate?: boolean;
  userUpdate?: boolean;
  userList?: boolean;

  search?: boolean;
}